requirements.txt
Django==2.2.5
gunicorn==19.9.0
whitenoise==4.1.3
psycopg2==2.8.3
pymysql==0.9.3
mysql-connector-python-rf==2.2.2
